Split test_snapshot.py into three files by source module
test_dev_repo_metrics.py - TestComputeMetricsDict
test_health_ranking.py - TestBandFromHealth, TestDictToRow, TestRankRows
test_snapshot.py - TestFromRepoPrs, TestBand, TestSummary, TestTeamAggregation, TestBuildSummary
